Remove Edit button
efibootmgr cannot actually edit an existing boot entry. That means that
editing must become a two step process, with deleting the old boot entry
first and creating a new one next. I'm not sure of the complications
that might arise from this yet, so better to remove the Edit button. If
the user wants to "edit" a boot entry, he can create a new one with the
desired settings and delete the old one.