Retain Enabled and BootNext after editing
"Editing" is actually a two-step process with creating a new entry and
deleting the old one, while making sure the new one retains all
attributes that haven't changed. So, we're making sure that if the
original boot entry is disabled or has BootNext set, these are also
applied in the new replacement entry.